pyproject.toml
[tool.poetry]
name = "cs50"
version = "0.1.0"
description = ""
authors = ["Michal Puškáč <133041741+michalpuskac@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
requests = "^2.32.3"
inflect = "^7.4.0"
pyfiglet = "^1.0.2"
emoji = "^2.14.0"
pytest = "^8.3.3"
tabulate = "^0.9.0"
validator-collection = "^1.5.0"
fpdf2 = "^2.8.1"
tox = "^4.23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
